--- a/.gitignore
+++ b/.gitignore
@@ -4,5 +4,6 @@ models/
 outputs/
 *.png
 *.jpg
+**/__pycache__
 
 **/.pytest_cache