## Related issues / PRs
- #7

## Summarize Changes
1. Added video driven source
2. Update defaults of the `crop_factor`, `retargeting_` and some parameters.
3. Removed meaningless inputs
4. Added video test along with changes
5. Autoacast to use precision for faster inference

Speed improvement as below


|    Inferencer      | Before | After | Speed Improvement |
|------------------|------------|-----------|------------------------|
| ExpressionEditor | 0.26 s       | 0.16 s     | 62%                 |
| AdvancedLivePortrait | 8.14 s      | 4.35 s     | 87%                 |