@jhj0517 2024-11-09
Inference is slower than official's. Probably torch autocast will fix it